@@ -4,7 +4,7 @@ version: 0.2.2
 homepage: https://github.com/ajomuch92/rolling-bottom-bar-flutter
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
